refactor(plugin-text): merge markdown shortcuts into use formatting options hook
